fix: 코드래빗 지적 반영 — aria-current·정렬필터 줄바꿈·상수 분리 #228
