+++ src/features/team/members/lib.ts
@@ -1,20 +1,16 @@
 import { ACTION_STATUS, isDelayed, MEMBER_STATUS } from "@/constants/domain";
+import {
+  TEAM_MEMBER_FILTER,
+  TEAM_MEMBER_FILTER_LABEL,
+  TEAM_MEMBER_SORT,
+  TEAM_MEMBER_SORT_LABEL,
+  type TeamMemberFilter,
+  type TeamMemberSort,
+} from "@/constants/team-member";
 
 import type { TeamMemberAction, TeamMemberStatusItem } from "./types";
 
-/** 정렬 3종(WORKFLOW.md §팀원 관리) — 값은 URL `?sort=`에 그대로 쓴다. */
-export const TEAM_MEMBER_SORT = {
-  ACTION_COUNT: "actionCount",
-  PROGRESS: "progress",
-  DELAYED: "delayed",
-} as const;
-export type TeamMemberSort = (typeof TEAM_MEMBER_SORT)[keyof typeof TEAM_MEMBER_SORT];
-
-export const TEAM_MEMBER_SORT_LABEL: Record<TeamMemberSort, string> = {
-  actionCount: "담당 액션 많은 순",
-  progress: "진척 낮은 순",
-  delayed: "지연 있는 순",
-};
+export type { TeamMemberFilter, TeamMemberSort };
 
 export const TEAM_MEMBER_SORT_TABS: { sort: TeamMemberSort; label: string }[] = [
   TEAM_MEMBER_SORT.ACTION_COUNT,
@@ -28,20 +24,6 @@ export const DEFAULT_TEAM_MEMBER_SORT: TeamMemberSort = TEAM_MEMBER_SORT.ACTION_
 export function parseTeamMemberSort(value: string | undefined): TeamMemberSort {
   return TEAM_MEMBER_SORT_TABS.find((t) => t.sort === value)?.sort ?? DEFAULT_TEAM_MEMBER_SORT;
 }
-
-/** 필터 3종 — 재직/휴직만 있다(퇴사자는 팀 로스터에서 이미 빠진 상태로 온다). */
-export const TEAM_MEMBER_FILTER = {
-  ALL: "all",
-  ACTIVE: "active",
-  VACATION: "vacation",
-} as const;
-export type TeamMemberFilter = (typeof TEAM_MEMBER_FILTER)[keyof typeof TEAM_MEMBER_FILTER];
-
-export const TEAM_MEMBER_FILTER_LABEL: Record<TeamMemberFilter, string> = {
-  all: "전체",
-  active: "재직",
-  vacation: "휴직",
-};
 
 export const TEAM_MEMBER_FILTER_TABS: { filter: TeamMemberFilter; label: string }[] = [
   TEAM_MEMBER_FILTER.ALL,
